Stored Procedure to create events
This stored procedure it's related to 3 different tables (events, artist_event, tickets). 
First I looked to insert data into the events table and then I set variables to create a relationship between the 3 tables. 
After we create the events, through this stored procedure, we immediately create tickets associated with it.